# servicios/nucleo-financiero/src/main/resources/openapi/nucleo-financiero.yaml
...
openapi: 3.1.0
info:
  title: nucleo-financiero
  version: 1.0.0
  description: >-
    Servicio nucleo-financiero. Un caso de uso = una operacion, con el codigo CU en el
    operationId. Los importes viajan como CADENA decimal, nunca como number.
servers:
  - url: /api/v1
paths:
...
  /billetera/retiros:
    post:
      operationId: solicitarRetiro             # CU-11
      summary: Solicitar un retiro hacia una cuenta de destino
      description: >-
        Retencion primero, pago despues, nunca al reves. Si se instruyera el pago
        antes de retener, entre las dos cosas la persona podria gastar el mismo
        saldo en otra operacion y el retiro saldria contra un disponible que ya no
        existe. El proveedor se instruye FUERA de la transaccion (invariante 6).
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRetiro' }
      responses:
        '201':
          description: Orden creada y saldo ya retenido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaRetiro' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU11-01 SALDO_INSUFICIENTE          · AP-CU11-02 MFA_REQUERIDO
          # AP-CU11-03 INSTRUMENTO_EN_ENFRIAMIENTO · AP-CU11-04 TITULAR_NO_COINCIDE
          # AP-CU11-05 BLOQUEO_DE_AUTORIDAD        · AP-CU11-06 ENCAJE_INCUMPLIDO
          # AP-CU11-07 MFA_INVALIDO (H2: evidencia provista pero no valida)
          # AP-CU11-13 la Idempotency-Key ya identifica otro retiro (otro importe, destino o cotizacion)
        '401':
          $ref: '#/components/responses/NoAutenticado'
...
components:
  parameters:
    ClaveIdempotencia:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Toda operacion con efecto la exige. La red duplica; reintentar tiene que ser
        seguro, y la clave se valida ANTES de escribir (invariante 7).
      schema: { type: string, format: uuid }
  schemas:
...
    EntradaRetiro:
      type: object
      additionalProperties: false
      required: [cuentaBilleteraId, monto, instrumentoDestinoId]
      properties:
        cuentaBilleteraId:    { type: string, format: uuid }
        monto:                { $ref: '#/components/schemas/Dinero' }
        instrumentoDestinoId: { type: string, format: uuid }
        # H2.S2.M5: evidenciaMfa reemplaza a factorMfa (JWT de evidencia step-up de
        # `identidad`, no un OTP corto). factorMfa se acepta hasta 2026-12-31 por
        # compatibilidad — ninguno de los dos es "required" a nivel de esquema
        # porque el rechazo por ausencia (MFA_REQUERIDO) lo decide el caso de uso,
        # no el contrato: es el mismo AP-CU11-02 tanto si falta el campo como si
        # llega vacio.
        evidenciaMfa: { type: string, minLength: 20, maxLength: 4096 }
        factorMfa:    { type: string, minLength: 6, maxLength: 8, deprecated: true }
        cotizacionId:
          type: string
          format: uuid
          description: >-
            La cotizacion que la persona vio y acepto (POST /billetera/cotizaciones, con la MISMA
            Idempotency-Key). Obligatoria mientras `aportaya.cotizacion.exigir-previa` este activa.
...
    SalidaRetiro:
      type: object
      additionalProperties: false
      required: [ordenRetiroId, estado, costoRetiro, montoNeto, retencionId]
      properties:
        ordenRetiroId: { type: string, format: uuid }
        estado:        { type: string, enum: [PENDIENTE, EN_REVISION, AUTORIZADA, PAGADA, RECHAZADA] }
        costoRetiro:   { $ref: '#/components/schemas/Dinero' }
        montoNeto:     { $ref: '#/components/schemas/Dinero' }
        retencionId:   { type: string, format: uuid }
...
    Dinero:
      type: object
      additionalProperties: false
      required: [monto, moneda]
      properties:
        monto:
          type: string
          pattern: '^-?\d+\.\d{2}$'     # CADENA: un number JSON es un doble
        moneda:
          type: string
          enum: [BOB, USD]
    Error:
      type: object
      additionalProperties: false
      required: [codigo, mensaje, trazaId]
      properties:
        codigo:   { type: string, pattern: '^AP-CU\d+-\d{2}$' }
        mensaje:  { type: string }
        detalle:  { type: object }
        trazaId:  { type: string }
...
  responses:
    ReglaDeNegocio:
      description: Regla de negocio de la aplicacion (422, no 400)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NoAutenticado:
      description: Sin sesion valida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }